Move InstanceInfo into a separate module
